# docs/swagger.yaml
definitions:
...
  LoginRequest:
    description: Fields untuk request login
    properties:
      email:
        description: Email user
        type: string
      password:
        description: Password user
        type: string
    required:
    - email
    - password
    type: object
...
  ResponseError:
    description: menampilkan response error dengan pesan dan detail, bisa berupa error
      dari binding atau error lainnya jika detail tidak ada, maka akan diabaikan
    properties:
      detail: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  ResponseSuccess:
    description: menampilkan response sukses hanya dengan pesan
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  ResponseSuccessToken:
    description: menampilkan response sukses dengan token biasanya digunakan untuk
      otentikasi atau otorisasi
    properties:
      success:
        type: boolean
      token:
        type: string
    type: object
  ResponseSuccessWithData:
    description: menampilkan response sukses dengan pesan dan data bisa berupa struct,
      slice, atau map jika data tidak ada, maka akan diabaikan
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  Users:
    description: Fields untuk user
    properties:
      created_at:
        description: Waktu pembuatan user
        type: string
      email:
        description: Email user
        type: string
      id:
        description: Id dari user
        type: integer
      name:
        description: Nama user
        type: string
      password:
        description: Password user
        type: string
      role:
        description: Peran user (admin/user/organizer) jika kosong berarti perannya
          user
        type: string
    required:
    - email
    - name
    - password
    type: object
info:
  contact: {}
paths:
...
  /api/users/login:
    post:
      consumes:
      - application/json
      description: Endpoint untuk login user
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil login
          schema:
            $ref: '#/definitions/ResponseSuccessToken'
        "400":
          description: Gagal login
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Login User
      tags:
      - User
  /api/users/me:
    get:
      consumes:
      - application/json
      description: Mengambil data user yang sedang login
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendapatkan user
          schema:
            $ref: '#/definitions/ResponseSuccessWithData'
        "400":
          description: Gagal mendapatkan user
          schema:
            $ref: '#/definitions/ResponseError'
      security:
      - BearerAuth: []
      summary: Get User
      tags:
      - User
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Endpoint untuk mendaftarkan user baru
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/Users'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mendaftarkan user
          schema:
            $ref: '#/definitions/ResponseSuccess'
        "400":
          description: Gagal mendaftarkan user
          schema:
            $ref: '#/definitions/ResponseError'
      summary: Register User
      tags:
      - User
swagger: "2.0"